// apps/thinkx/src/app/shared/feature-type.enum.ts
export enum FeatureType {
  STATUS = 'status',
  TOGGLE = 'toggle',
  DROPDOWN = 'dropdown',
  TIMER = 'timer',
  REMINDER = 'reminder',
  TOGGLETIMER = 'toggle + timer',
  STATUSTEXTFIELD = 'status + text-field',
  TOGGLEDROPDOWN = 'toggle + dropdown',
  TOGGLEREMINDER = 'toggle + reminder',
  MASTERRESET = 'master reset',
  CHANGEPASSWORD= 'change password',
}